ctmap: add mock maps and unit test for BPF conntrack JSON dump
Add mock maps to the ctmap package. They can be used for unit tests.

Speaking of: Add a unit test for JSON dump of the conntrack maps via the
cli. We dump and parse JSON, then we check that some of the values are
as expected, based on what we inserted into the mock maps..

Signed-off-by: Quentin Monnet <quentin@isovalent.com>