ipsec: ensure to not remove the wrong xfrm states
This ensures that in case the SPI in use has changed, don't remove any
states at all. When the in use SPI has changed, there is already another
gorutine that is waiting to do the cleanup, so it will eventually be
done.

This will just limit the race window, from a minutes long to a lot less.
We should probably also grab the _ipSecKeysGlobalLock_ as well as
comparing the content of the ipSecKeysGlobal map to reduce the race to
zero; unless other refactors are done.

Signed-off-by: Odin Ugedal <ougedal@palantir.com>