labels(poc): Reduce overhead of label comparison for k8s
We know that out of say 40k rules on a large cluster, a tiny number does match. So its all about making the match return false first. So first of all, we see that the first label will always be the same, then the second will almost always be the same. So we reverse, since we know that the uid is distinct per policy. That means that for a non-match (aka. all of them), we reduce the number of comparisons by reversing the check order, and starting with the UID part.

This ^ is true for k8s, since that is the order of the labels in that
case.

Also, we know that for that uid, its the actual value that is different, not the key. So we start comparing the value first, and not the source, then key, then value.

Overall, this should reduce the number of string comparisons from ~number of policy rules * 8 to ~number of policy rules * 1 if my maths and brain is correct.

note: this is a fix for our internal issue, so it might not make sense
for upstream.